+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/trace/interceptor/M_HU_Trx_Hdr.java
@@ -2,11 +2,14 @@ package de.metas.handlingunits.trace.interceptor;
 
 import java.util.List;
 
+
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
 import org.adempiere.util.Services;
 import org.compiere.Adempiere;
 import org.compiere.model.ModelValidator;
+
+import com.google.common.annotations.VisibleForTesting;
 
 import de.metas.handlingunits.IHUTrxDAO;
 import de.metas.handlingunits.model.I_M_HU_Trx_Hdr;
@@ -39,11 +42,29 @@ import lombok.NonNull;
 @Interceptor(I_M_HU_Trx_Hdr.class)
 public class M_HU_Trx_Hdr
 {
-	@ModelChange(timings =
-		{
-				ModelValidator.TYPE_AFTER_CHANGE,
-				ModelValidator.TYPE_AFTER_NEW
-		}, ifColumnsChanged = I_M_HU_Trx_Hdr.COLUMNNAME_Processed, afterCommit = true)
+
+	private final HUTraceEventsCreateAndAdd huTraceEventsCreateAndAdd;
+
+	public M_HU_Trx_Hdr()
+	{
+		huTraceEventsCreateAndAdd = null; // will take the value from spring
+	}
+
+	/**
+	 * To be used in testing.
+	 * 
+	 * @param huTraceEventsCreateAndAdd
+	 */
+	@VisibleForTesting
+	public M_HU_Trx_Hdr(@NonNull final HUTraceEventsCreateAndAdd huTraceEventsCreateAndAdd)
+	{
+		this.huTraceEventsCreateAndAdd = huTraceEventsCreateAndAdd;
+	}
+
+	@ModelChange(timings = {
+			ModelValidator.TYPE_AFTER_CHANGE,
+			ModelValidator.TYPE_AFTER_NEW
+	}, ifColumnsChanged = I_M_HU_Trx_Hdr.COLUMNNAME_Processed, afterCommit = true)
 	public void addTraceEvent(@NonNull final I_M_HU_Trx_Hdr huTrxHeader)
 	{
 		if (!huTrxHeader.isProcessed())
@@ -53,8 +74,19 @@ public class M_HU_Trx_Hdr
 
 		final IHUTrxDAO huTrxDAO = Services.get(IHUTrxDAO.class);
 		final List<I_M_HU_Trx_Line> huTrxLines = huTrxDAO.retrieveTrxLines(huTrxHeader);
-		
-		final HUTraceEventsCreateAndAdd huTraceEventsCreateAndAdd = Adempiere.getBean(HUTraceEventsCreateAndAdd.class);
-		huTraceEventsCreateAndAdd.createAndAddFor(huTrxHeader, huTrxLines.stream());
+
+		final HUTraceEventsCreateAndAdd huTraceEventsCreateAndAdd = getHuTraceEventsCreateAndAdd();
+		huTraceEventsCreateAndAdd.createAndAddFor(huTrxHeader, huTrxLines);
 	}
+
+	
+	private HUTraceEventsCreateAndAdd getHuTraceEventsCreateAndAdd()
+	{
+		if (this.huTraceEventsCreateAndAdd != null)
+		{
+			return huTraceEventsCreateAndAdd;
+		}
+		return Adempiere.getBean(HUTraceEventsCreateAndAdd.class);
+	}
+
 }
